+++ CHANGELOG.md
@@ -12,8 +12,8 @@ Mail:
 
 System:
 
+* Backups now use duplicity's built-in gpg symmetric AES256 encryption rather than my home-brewed encryption. Old backups will be incorporated inside the first backup after this update but then deleted from disk (i.e. your backups from the previous few days will be backed up).
 * Internationalized Domain Names (IDNs) should now work in email. If you had custom DNS or custom web settings for internationalized domains, check that they are still working.
-
 
 v0.08 (April 1, 2015)
 ---------------------